ai_ml_track/requirements.txt
sentence-transformers
scikit-learn
google-generativeai
python-dotenvS